Use VirtualJSFiles instead of ResolvedJSDependency in the JSEnv API.
This removes the annoying dependency (sic!) of the `JSEnv` API on
`ResolvedJSDependency`. This was the only part of `tools.jsdep`
that was used in the `JSEnv` API, so it decouples those.

This will be very important as we separate `jsDependencies` in a
separate plugin, as the `tools.jsdep` package should go with that,
out of the core.