Fix #2874: Remove the distinction libs/code in the JSEnv API.
`JSEnv#jsRunner` and friends `asyncRunner` and `comRunner` now
only take a `Seq[VirtualJSFiles]`, containing all JS files to be
given to the JS environment.

The JS environments therefore make no distinction between "library"
files and the "code" file. The disinction was arbitrary anyway,
and unifying them makes sure that all the JS files are first-class.
In particular, it forced to solve shortcomings in terms of error
handling in `JSDOMNodeJSEnv`.